+++ .github/workflows/bika.yaml
@@ -24,7 +24,7 @@ jobs:
               run: pip install -r ./requirements.txt
 
             - name: "运行 哔咔漫画自动打哔咔"
-              run: python3 function/bika/bika.py
+              run: python3 function/bika.py
               env:
                   BIKA_USER: ${{ secrets.BIKA_USER }}
                   BIKA_PASS: ${{ secrets.BIKA_PASS }}
@@ -38,4 +38,4 @@ jobs:
                   BARK_SOUND: ${{ secrets.BARK_SOUND }}
                   DD_BOT_TOKEN: ${{ secrets.DD_BOT_TOKEN }}
                   DD_BOT_SECRET: ${{ secrets.DD_BOT_SECRET }}
-#最后推送 2021年4月09日
+#最后推送 2021年3月26日
