+++ .github/workflows/bilibili.yaml
@@ -24,7 +24,7 @@ jobs:
               run: pip install -r ./requirements.txt
 
             - name: "运行 哔哩哔哩签到"
-              run: python3 function/bilibili/bilibili.py
+              run: python3 function/bilibili.py
               env:
                   BILI_USER: ${{ secrets.BILI_USER }}
                   BILI_PASS: ${{ secrets.BILI_PASS }}
@@ -41,4 +41,4 @@ jobs:
                   BARK_SOUND: ${{ secrets.BARK_SOUND }}
                   DD_BOT_TOKEN: ${{ secrets.DD_BOT_TOKEN }}
                   DD_BOT_SECRET: ${{ secrets.DD_BOT_SECRET }}
-#最后推送 2021年4月09日
+#最后推送 2021年3月26日
