+++ .github/workflows/tieba.yaml
@@ -22,7 +22,7 @@ jobs:
                   npm install
             - name: "运行【百度签到】"
               run: |
-                  node function/tieba/tieba.js
+                  node function/tieba.js
               env:
                   BDUSS: ${{ secrets.BDUSS }}
                   JUMP: ${{ secrets.BDUSS-JUMP }}
@@ -41,4 +41,4 @@ jobs:
                   BARK_SOUND: ${{ secrets.BARK_SOUND }}
                   DD_BOT_TOKEN: ${{ secrets.DD_BOT_TOKEN }}
                   DD_BOT_SECRET: ${{ secrets.DD_BOT_SECRET }}
-#最后推送 2021年4月09日
+#最后推送 2021年3月26日
