+++ .github/workflows/v_video.yaml
@@ -22,7 +22,7 @@ jobs:
                   npm install
             - name: "运行【腾讯视频签到】"
               run: |
-                  node function/q_video/q_video.js
+                  node function/v_video.js
               env:
                   V_REF_URL: ${{ secrets.V_REF_URL }}
                   V_COOKIE: ${{ secrets.V_COOKIE }}
@@ -41,4 +41,4 @@ jobs:
                   BARK_SOUND: ${{ secrets.BARK_SOUND }}
                   DD_BOT_TOKEN: ${{ secrets.DD_BOT_TOKEN }}
                   DD_BOT_SECRET: ${{ secrets.DD_BOT_SECRET }}
-#最后推送 2021年4月09日
+#最后推送 2021年3月26日
