+++ function/bika.py
@@ -7,7 +7,7 @@ import string
 import time
 import os
 import sys;
-sys.path.append("My-Actions/function/bika/")
+sys.path.append("My-Actions/function/")
 from sendNotify import *
 from http import client
