+++ function/wps_invitation.py
@@ -4,13 +4,13 @@ import requests
 import time
 import json
 import sys
+sys.path.append("My-Actions/function/")
 import pytz
 import datetime
 import re
 import os
-from io import StringIO
-sys.path.append("My-Actions/function/wps")
 from sendNotify import *
+from io import StringIO
 
 # Python版本 3.6, 该脚本仅供分享交流和学习, 不允许用于任何非法途径, 否则后果自负, 作者对此不承担任何责任
 # 20210122更新: 添加WPS小程序会员群集结功能 (如需仅执行群集结功能, 请将执行方法由'index.main_handler'更改为'index.wps_massing' )
